+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 -----
 - Add support for the ``allow_custom_choice`` field property in Python
   processes
+- Add ordering by contributor's first and last name to Collection viewset
 
 Fixed
 -----